Reorganise test scenes into playlists specific namespace
