+++ osu.Game.Tests/Visual/Playlists/TestScenePlaylistsParticipantsList.cs
@@ -6,7 +6,7 @@ using osu.Framework.Graphics;
 using osu.Game.Screens.Multi.Components;
 using osu.Game.Users;
 
-namespace osu.Game.Tests.Visual.Multiplayer
+namespace osu.Game.Tests.Visual.Playlists
 {
     public class TestScenePlaylistsParticipantsList : MultiplayerTestScene
     {
